# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MClass Course Management API",
        "description": "API documentation for MClass course management system",
        "contact": {
            "email": "support@mclass.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "MClass API Server"
        }
    ],
    "paths": {
        "/api/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get all members",
                "description": "Retrieve a list of all members with their profiles, contacts, and backgrounds",
                "operationId": "getMembersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nickname": {
                                                "type": "string",
                                                "example": "JohnDoe"
                                            },
                                            "account": {
                                                "type": "string",
                                                "example": "john.doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "john@example.com"
                                            },
                                            "status": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Create a new member",
                "description": "Create a new member with profile, contact, and background information",
                "operationId": "createMember",
                "requestBody": {
                    "description": "Member data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member",
                                    "profile",
                                    "contact",
                                    "background"
                                ],
                                "properties": {
                                    "member": {
                                        "properties": {
                                            "nickname": {
                                                "type": "string",
                                                "example": "JohnDoe"
                                            },
                                            "account": {
                                                "type": "string",
                                                "example": "john.doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "john@example.com"
                                            },
                                            "email_valid": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "password123"
                                            },
                                            "status": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "profile": {
                                        "properties": {
                                            "lastname": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "firstname": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "gender": {
                                                "type": "string",
                                                "enum": [
                                                    "male",
                                                    "female",
                                                    "other"
                                                ],
                                                "example": "male"
                                            },
                                            "birthday": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1990-01-01"
                                            },
                                            "job": {
                                                "type": "string",
                                                "example": "Software Engineer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nickname": {
                                            "type": "string",
                                            "example": "JohnDoe"
                                        },
                                        "account": {
                                            "type": "string",
                                            "example": "john.doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products (courses)",
                "description": "Retrieve a list of all course products with their information, followers, and visible students",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "club_course_info_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 99.99
                                            },
                                            "inventory": {
                                                "type": "integer",
                                                "example": 50
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Retrieve detailed information about a specific course product",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "club_course_info_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "inventory": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Members",
            "description": "Members"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}